Share code between validate{Paint,Layout}Property
